Done Coding

+++ src/main/java/com/crud/java_crud_mysql/HelloController.java
@@ -2,10 +2,7 @@ package com.crud.java_crud_mysql;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.Alert;
-import javafx.scene.control.ButtonType;
-import javafx.scene.control.Label;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
@@ -33,6 +30,9 @@ public class HelloController {
 
     @FXML
     private TextField Dlocation_DL;
+
+    @FXML
+    private TextArea Dlocation_List;
 
     @FXML
     private Label info;
@@ -86,6 +86,7 @@ public class HelloController {
         // clear Dnumber_DL, Dlocation_DL
         Dnumber_DL.setText("");
         Dlocation_DL.setText("");
+        Dlocation_List.setText("");
     }
 
     public void onInsertClick(ActionEvent actionEvent) {
@@ -105,7 +106,7 @@ public class HelloController {
                     if (isInt(Mgr_ssn, "Please input number for Mgr_ssn") && isValidSsn("Please input 9 digits for Mgr_ssn")) {
                         if (isDate(Mgr_start_due, "Please input valid date for Mgr_start_due")) {
                             // Find if the Dnumber is already in the database
-                            if (!isDnumberExist(Dnumber.getText())) {
+                            if (isDnumberExist(Dnumber.getText())) {
                                 info.setText("Info : Error The Dnumber is already in the database. Try another number.");
                             } else {
                                 statement.executeUpdate("INSERT INTO department VALUES ('" + Dname.getText() + "','" + Integer.parseInt(Dnumber.getText()) + "','" + Integer.parseInt(Mgr_ssn.getText()) + "','" + Mgr_start_due.getText() + "')");
@@ -136,6 +137,20 @@ public class HelloController {
         return false;
     }
 
+    private boolean isDLnumberExist(String text) {
+        try {
+            Connection connection = DBConnect();
+            Statement statement = connection.createStatement();
+            ResultSet resultSet = statement.executeQuery("SELECT * FROM dept_location WHERE Dnumber = " + text);
+            if (resultSet.next()) {
+                return true;
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return false;
+    }
+
     private boolean isValidSsn(String msg) {
         if (Mgr_ssn.getText().length() == 9) {
             return true;
@@ -149,7 +164,7 @@ public class HelloController {
         try {
             Connection connection = DBConnect();
             Statement statement = connection.createStatement();
-            if (Dnumber_DL.getText().isEmpty()) {
+            if (Dnumber.getText().isEmpty()) {
                 info.setText("Info : Please enter a DNumber to be updated");
             } else if (Dname.getText().isEmpty()) {
                 info.setText("Info : Please enter a DName");
@@ -158,11 +173,15 @@ public class HelloController {
             } else if (Mgr_start_due.getText().isEmpty()) {
                 info.setText("Info : Please enter a Mgr_start_due");
             } else {
-                if (isInt(Dnumber_DL, "Please input number for Dnumber")) {
-                    if (isInt(Mgr_ssn, "Please input number for Mgr_ssn") && isValidSsn("Please input 9 digits for Mgr_ssn")) {
-                        if (isDate(Mgr_start_due, "Please input valid date for Mgr_start_due")) {
-                            statement.executeUpdate("UPDATE department SET Dname = '" + Dname.getText() + "', Dnumber = '" + Integer.parseInt(Dnumber.getText()) + "', Mgr_ssn = '" + Integer.parseInt(Mgr_ssn.getText()) + "', Mgr_start_due = '" + Mgr_start_due.getText() + "' WHERE Dnumber = '" + Integer.parseInt(Dnumber.getText()) + "'");
-                            info.setText("Info : Updated successfully");
+                if (!isDnumberExist(Dnumber.getText())) {
+                    info.setText("Info : Error The Dnumber is not in the database. Try another number.");
+                } else {
+                    if (isInt(Dnumber, "Please input number for Dnumber")) {
+                        if (isInt(Mgr_ssn, "Please input number for Mgr_ssn") && isValidSsn("Please input 9 digits for Mgr_ssn")) {
+                            if (isDate(Mgr_start_due, "Please input valid date for Mgr_start_due")) {
+                                statement.executeUpdate("UPDATE department SET Dname = '" + Dname.getText() + "', Dnumber = '" + Integer.parseInt(Dnumber.getText()) + "', Mgr_ssn = '" + Integer.parseInt(Mgr_ssn.getText()) + "', Mgr_start_due = '" + Mgr_start_due.getText() + "' WHERE Dnumber = '" + Integer.parseInt(Dnumber.getText()) + "'");
+                                info.setText("Info : Updated successfully");
+                            }
                         }
                     }
                 }
@@ -178,14 +197,20 @@ public class HelloController {
         try {
             Connection connection = DBConnect();
             Statement statement = connection.createStatement();
-            if (Dnumber_DL.getText().isEmpty()) {
+            if (Dnumber.getText().isEmpty()) {
                 info.setText("Info : Please enter a DNumber to be deleted");
             } else {
-                if (isInt(Dnumber_DL, "Please input number for Dnumber")) {
-                    statement.executeUpdate("DELETE FROM department WHERE Dnumber = '" + Integer.parseInt(Dnumber.getText()) + "'");
-                    info.setText("Info : Deleted successfully");
-                    // Clear the Fields
-                    clearD();
+                if (!isDnumberExist(Dnumber.getText())) {
+                    info.setText("Info : Error The Dnumber is not in the database. Try another number.");
+                } else if (isDLnumberExist(Dnumber.getText())) {
+                    info.setText("Info : Error The Dnumber is in the dept_location table. Try after deleting that entry.");
+                } else {
+                    if (isInt(Dnumber, "Please input number for Dnumber")) {
+                        statement.executeUpdate("DELETE FROM department WHERE Dnumber = '" + Integer.parseInt(Dnumber.getText()) + "'");
+                        info.setText("Info : Deleted successfully");
+                        // Clear the Fields
+                        clearD();
+                    }
                 }
             }
             System.out.println("Deleted");
@@ -202,6 +227,7 @@ public class HelloController {
             info.setText("Info : ");
             // Clear the Fields
             Dlocation_DL.setText("");
+            Dlocation_List.setText("");
 
             if (Dnumber_DL.getText().isEmpty()) {
                 info.setText("Info : Please enter a DNumber to be searched");
@@ -209,16 +235,27 @@ public class HelloController {
                 if (isInt(Dnumber_DL, "Please input integer for Dnumber")) {
                     ResultSet resultSet = statement.executeQuery("SELECT * FROM dept_location WHERE Dnumber = " + Integer.parseInt(Dnumber_DL.getText()));
                     if (resultSet.next()) {
+                        // set Dlocation_DL to ""
+                        Dlocation_DL.setText("");
+                        Dlocation_List.setText("");
+
+                        System.out.println(resultSet.getString("Dlocation"));
+                        Dnumber_DL.setText(resultSet.getString("Dnumber"));
+                        // append Dlocation with resultSet.getString("Dlocation")
+                        Dlocation_List.setText(Dlocation_DL.getText() + resultSet.getString("Dlocation") + "\n");
                         Dlocation_DL.setText(resultSet.getString("Dlocation"));
+
+                        while (resultSet.next()) {
+                            System.out.println(resultSet.getString("Dlocation"));
+                            Dnumber_DL.setText(resultSet.getString("Dnumber"));
+                            // append Dlocation with resultSet.getString("Dlocation")
+                            Dlocation_List.setText(Dlocation_List.getText() + resultSet.getString("Dlocation") + "\n");
+                        }
                     } else {
                         info.setText("Info : No such Dnumber in the database");
                     }
-//                    while (resultSet.next()) {
-//                        Dnumber_DL.setText(resultSet.getString("Dnumber"));
-//                        Dlocation_DL.setText(resultSet.getString("Dlocation"));
-//                    }
-                }
-                else {
+
+                } else {
                     info.setText("Info : Please enter a DNumber to be searched");
                 }
             }
@@ -240,8 +277,7 @@ public class HelloController {
                 // if Dnumber is not in the database
                 if (!isDnumberExist(Dnumber_DL.getText())) {
                     info.setText("Info : Error The Dnumber is not in the database. Try another number.");
-                }
-                else {
+                } else {
                     if (isInt(Dnumber_DL, "Please input number for Dnumber")) {
                         statement.executeUpdate("INSERT INTO dept_location VALUES ('" + Integer.parseInt(Dnumber_DL.getText()) + "','" + Dlocation_DL.getText() + "')");
                         info.setText("Info : Inserted successfully");
@@ -264,10 +300,15 @@ public class HelloController {
             } else if (Dlocation_DL.getText().isEmpty()) {
                 info.setText("Info : Please enter a Dlocation");
             } else {
-                if (isInt(Dnumber_DL, "Please input number for Dnumber")) {
-                    statement.executeUpdate("UPDATE dept_location SET Dnumber = '" + Integer.parseInt(Dnumber_DL.getText()) + "', Dlocation = '" + Dlocation_DL.getText() + "' WHERE Dnumber = '" + Integer.parseInt(Dnumber_DL.getText()) + "'");
-                    info.setText("Info : Updated successfully");
+                if (!isDnumberExist(Dnumber_DL.getText()) && !isDLnumberExist(Dnumber_DL.getText())) {
+                    info.setText("Info : Error The Dnumber is not in the database. Try another number.");
+                } else {
+                    if (isInt(Dnumber_DL, "Please input number for Dnumber")) {
+                        statement.executeUpdate("UPDATE dept_location SET Dnumber = '" + Integer.parseInt(Dnumber_DL.getText()) + "', Dlocation = '" + Dlocation_DL.getText() + "' WHERE Dnumber = '" + Integer.parseInt(Dnumber_DL.getText()) + "'");
+                        info.setText("Info : Updated successfully");
+                    }
                 }
+
             }
             //statement.executeUpdate("UPDATE dept_location SET Dnumber = '" + Integer.parseInt(Dnumber_DL.getText()) + "', Dlocation = '" + Dlocation_DL.getText() + "' WHERE Dnumber = '" + Integer.parseInt(Dnumber_DL.getText()) + "'");
             System.out.println("Updated");
@@ -283,10 +324,14 @@ public class HelloController {
             if (Dnumber_DL.getText().isEmpty()) {
                 info.setText("Info : Please enter a DNumber to be deleted");
             } else {
-                if (isInt(Dnumber_DL, "Please input number for Dnumber")) {
-                    statement.executeUpdate("DELETE FROM dept_location WHERE Dnumber = '" + Integer.parseInt(Dnumber_DL.getText()) + "'");
-                    info.setText("Info : Deleted successfully");
-                    clearDL();
+                if (!isDLnumberExist(Dnumber_DL.getText())) {
+                    info.setText("Info : Error The Dnumber is not in the database. Try another number.");
+                } else {
+                    if (isInt(Dnumber_DL, "Please input number for Dnumber")) {
+                        statement.executeUpdate("DELETE FROM dept_location WHERE Dnumber = '" + Integer.parseInt(Dnumber_DL.getText()) + "'");
+                        info.setText("Info : Deleted successfully");
+                        clearDL();
+                    }
                 }
             }
             //statement.executeUpdate("DELETE FROM dept_location WHERE Dnumber = '" + Integer.parseInt(Dnumber_DL.getText()) + "'");
